Agrego controller, endpoint post y getId

+++ FarmaciaVerifarmaChallenge/Domain/Entities/Farmacia.cs
@@ -8,8 +8,8 @@ namespace FarmaciaVerifarmaChallenge.Domain.Entities
         public string Nombre { get; set; }
         public string Direccion { get; set; }
         [Column(TypeName = "decimal(8, 5)")] //maximo num posible 90 + 5 digitos + digito para negativo = 8
-        public int Latitud { get; set; }
+        public decimal Latitud { get; set; }
         [Column(TypeName = "decimal(9, 5)")] //maximo num posible 180
-        public int Longitud { get; set; }
+        public decimal Longitud { get; set; }
     }
 }
